Move edit to a separate file
