+++ includes/block-editor/index.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=6)}([function(t,e){!function(){t.exports=this.wp.element}()},function(t,e){!function(){t.exports=this.wp.i18n}()},function(t,e,n){var r=n(7),o=n(8),i=n(9),c=n(11);t.exports=function(t,e){return r(t)||o(t,e)||i(t,e)||c()}},function(t,e){!function(){t.exports=this.wp.apiFetch}()},function(t,e){!function(){t.exports=this.wp.blocks}()},function(t,e){!function(){t.exports=this.wp.components}()},function(t,e,n){"use strict";n.r(e);var r=n(2),o=n.n(r),i=n(0),c=n(3),u=n.n(c),a=n(1),f=n(4),l=n(5),s=new Map;Object(f.registerBlockType)("contact-form-7/contact-form-selector",{title:Object(a.__)("Contact Form 7","contact-form-7"),description:Object(a.__)("Insert a contact form you have created with Contact Form 7.","contact-form-7"),icon:"email",category:"widgets",attributes:{id:{type:"integer"},title:{type:"string"}},edit:function(t){var e=t.attributes,n=t.setAttributes,r=Array.from(s.values(),(function(t){return{value:t.id,label:t.title}}));if(!e.id){var o=r[0];n({id:parseInt(o.value),title:o.label})}return Object(i.createElement)(l.SelectControl,{options:r,value:e.id,onChange:function(t){return n({id:parseInt(t),title:s.get(parseInt(t)).title})}})},save:function(t){var e=t.attributes;return Object(i.createElement)("div",null,'[contact-form-7 id="',e.id,'" title="',e.title,'"]')}}),u()({path:"contact-form-7/v1/contact-forms"}).then((function(t){Object.entries(t).forEach((function(t){var e=o()(t,2),n=(e[0],e[1]);s.set(n.id,n)}))}))},function(t,e){t.exports=function(t){if(Array.isArray(t))return t}},function(t,e){t.exports=function(t,e){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t)){var n=[],r=!0,o=!1,i=void 0;try{for(var c,u=t[Symbol.iterator]();!(r=(c=u.next()).done)&&(n.push(c.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==u.return||u.return()}finally{if(o)throw i}}return n}}},function(t,e,n){var r=n(10);t.exports=function(t,e){if(t){if("string"==typeof t)return r(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?r(t,e):void 0}}},function(t,e){t.exports=function(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}}]);
+!function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=11)}([function(t,e){!function(){t.exports=this.wp.element}()},function(t,e){!function(){t.exports=this.wp.i18n}()},function(t,e){!function(){t.exports=this.wp.blocks}()},function(t,e,n){var r=n(6),o=n(7),i=n(8),c=n(10);t.exports=function(t,e){return r(t)||o(t,e)||i(t,e)||c()}},function(t,e){!function(){t.exports=this.wp.apiFetch}()},function(t,e){!function(){t.exports=this.wp.components}()},function(t,e){t.exports=function(t){if(Array.isArray(t))return t}},function(t,e){t.exports=function(t,e){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t)){var n=[],r=!0,o=!1,i=void 0;try{for(var c,u=t[Symbol.iterator]();!(r=(c=u.next()).done)&&(n.push(c.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==u.return||u.return()}finally{if(o)throw i}}return n}}},function(t,e,n){var r=n(9);t.exports=function(t,e){if(t){if("string"==typeof t)return r(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?r(t,e):void 0}}},function(t,e){t.exports=function(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}},function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}},function(t,e,n){"use strict";n.r(e);var r=n(0),o=n(1),i=n(2),c=n(3),u=n.n(c),a=n(4),f=n.n(a),l=n(5),s=new Map;f()({path:"contact-form-7/v1/contact-forms"}).then((function(t){Object.entries(t).forEach((function(t){var e=u()(t,2),n=(e[0],e[1]);s.set(n.id,n)}))})),Object(i.registerBlockType)("contact-form-7/contact-form-selector",{title:Object(o.__)("Contact Form 7","contact-form-7"),description:Object(o.__)("Insert a contact form you have created with Contact Form 7.","contact-form-7"),icon:"email",category:"widgets",attributes:{id:{type:"integer"},title:{type:"string"}},edit:function(t){var e=t.attributes,n=t.setAttributes,o=Array.from(s.values(),(function(t){return{value:t.id,label:t.title}}));if(!e.id){var i=o[0];n({id:parseInt(i.value),title:i.label})}return Object(r.createElement)(l.SelectControl,{options:o,value:e.id,onChange:function(t){return n({id:parseInt(t),title:s.get(parseInt(t)).title})}})},save:function(t){var e=t.attributes;return Object(r.createElement)("div",null,'[contact-form-7 id="',e.id,'" title="',e.title,'"]')}})}]);
