+++ includes/block-editor/src/index.js
@@ -1,9 +1,7 @@
-import apiFetch from '@wordpress/api-fetch';
 import { __ } from '@wordpress/i18n';
 import { registerBlockType } from '@wordpress/blocks';
-import { SelectControl } from '@wordpress/components';
 
-const contactForms = new Map();
+import edit from './edit';
 
 registerBlockType( 'contact-form-7/contact-form-selector', {
 
@@ -24,33 +22,7 @@ registerBlockType( 'contact-form-7/contact-form-selector', {
 		},
 	},
 
-	edit: ( { attributes, setAttributes } ) => {
-		const options = Array.from( contactForms.values(), ( val ) => { 
-			return { value: val.id, label: val.title };
-		} );
-
-		if ( ! attributes.id ) {
-			const firstOption = options[0];
-
-			setAttributes( {
-				id: parseInt( firstOption.value ),
-				title: firstOption.label,
-			} );
-		}
-
-		return(
-			<SelectControl
-				options={ options }
-				value={ attributes.id }
-				onChange={
-					( value ) => setAttributes( {
-						id: parseInt( value ),
-						title: contactForms.get( parseInt( value ) ).title
-					} )
-				}
-			/>
-		);
-	},
+	edit,
 
 	save: ( { attributes } ) => {
 		return(
@@ -59,12 +31,4 @@ registerBlockType( 'contact-form-7/contact-form-selector', {
 			</div>
 		);
 	},
-} );
-
-apiFetch( {
-	path: 'contact-form-7/v1/contact-forms',
-} ).then( response => {
-	Object.entries( response ).forEach( ( [ key, value ] ) => {
-		contactForms.set( value.id, value );
-	} );
 } );
